OPSAPS-51686. Pass cookies for CM UI when configured with HA
We are currently deploying all services (including CM) using HA enabled,
even though there's only one endpoint. That means that the
"ha-classname" is used for dispatch instead of the regular "classname".
This caused the ConfigurableDispatch that was set for CM to be ignored,
and instead DefaultDispatch was used. DefaultDispatch drops Set-Cookie
headers, which made CM sessions break.

Tested by manually updating the service.xml in a deployed cluster and
verifying that I got a cookie in the end-user response. I also verified
that I can now go through the "add service" wizard, the original issue
noted by OPSAPS-51633.

Change-Id: Ia1fbc1d6fd541092a69c92df1ca159c67376cae3