[cm] OPSAPS-51592 Remove Referer rewrite rule in CM-UI
There is a rewrite rule in CM-UI that aims to replace the host/port in Referer
headers with that of the CM server host. This is to avoid failing CM's CSRF
Referer check. However, the rule is overzealous and allows a request from any
host to be accepted by CM, effectively preventing the CSRF Referer check from
working.

The plan is to change CM to accept a frontend URL which the CSRF Referer check
can compare against. That config is meant to be set to the Knox host, allowing
requests proxied through Knox to pass the CSRF check. That means this rewrite
rule is no longer necessary.

It is also believed that this rewrite rule is also affecting request bodies and
not just the header, causing other sorts of mischief.

Testing:
* I have a CM build with the new frontend URL change, and along with this
change, I verified that requests to the CM UI proxied through Knox passed the
CSRF Referer check and succeeded.

Change-Id: I2bf6f650b08698ec3d4933fedff7d14bc9d033b3