CDPD-10830 - Removed topology name filtering when processing CM descriptors
The topology name filtering was added - and left there mistakenly - before I added the check to save a descriptor only if it was actually changed. This filter no longer needs because even if we parse the entire CM descriptor the descriptors/providers that are not changed won't be persisted.

Tested manually in my cluster:
- added/modified parameters in cdp-proxy and cdp-proxy-api descriptors (Knox Simplified Topology Management)
- added/modified a custom descriptor
- enabled/disabled services in cdp-proxy and cdp-proxy-api topologies via the CM UI

All was working as expected.

Change-Id: I6995d14036caad514e2edf0d879d2708a67304c4