CDPD-1202 - End-users can customize failover/retry attributes.
In case of request execution errors, the client tried to failover to an active IDBroker server host (if there was any) or re-tried to execute the request on the same IDBRoker host. The number of failover/retries were hardcoded in the Java code as well as the sleep time between those actions. From now on, there is a way to customize them through core-site.xml.

Change-Id: Icf6436c34d52ceb4a303d6cbddcad7cdf424ddae