CDPD-28900 - Make sure binding code sends correct host header after failover
This patch adds Host header to the cab session which is then added to the outgoing IDBroker request. We don't have a way to directly modify the Request object given how it is set up (cab session is part of Request and Request object does not expose methods to modify request url).

This patch was tested on the test cluster.

Change-Id: I94b7a0606f03afe632c7a04e42081169b4d0ea8f