CDPD-55580: PollingConfigurationAnalyzer ignores services without service model generator (#755)
Since WebHCat and HiveMetastore are not supported in CDP, their ServiceModelGenerator resources are inactivated. Therefore the new logic will not trigger a topology redeployment when the HIVE-1 service is restarted (please note this does not have any effect on our HIVE operations, the HiveServer2 instances are runnig under HIVE_ON_TEZ).

Change-Id: Ide368615b25e47a34ad536a4caa795dd2297ea10